feat: use extracted list component to show vms
