--- src/store/slices/test.json
@@ -1,135 +1,0 @@
-{
-    "apiVersion": "pool.kubevirt.io/v1alpha1",
-    "kind": "VirtualMachinePool",
-    "metadata": {
-      "name": "test",
-      "namespace": "default",
-      "labels": {
-        "kubevirt.io/vmpool": "test",
-        "kubevirt-manager.io/managed": "true",
-        "fw.kubevirt-manager.io/pool-name": "test",
-        "cloud-init.kubevirt-manager.io/username": ""
-      }
-    },
-    "spec": {
-      "replicas": 2,
-      "selector": {
-        "matchLabels": {
-          "kubevirt.io/vmpool": "test",
-          "kubevirt-manager.io/managed": "true",
-          "fw.kubevirt-manager.io/pool-name": "test",
-          "cloud-init.kubevirt-manager.io/username": ""
-        }
-      },
-      "virtualMachineTemplate": {
-        "metadata": {
-          "labels": {
-            "kubevirt.io/vmpool": "test",
-            "kubevirt-manager.io/managed": "true",
-            "fw.kubevirt-manager.io/pool-name": "test",
-            "cloud-init.kubevirt-manager.io/username": ""
-          }
-        },
-        "spec": {
-          "dataVolumeTemplates": [
-            {
-              "apiVersion": "cdi.kubevirt.io/v1beta1",
-              "kind": "DataVolume",
-              "metadata": {
-                "name": "default-test-disk1",
-                "namespace": "default",
-                "annotations": {
-                  "cdi.kubevirt.io/storage.deleteAfterCompletion": "false"
-                }
-              },
-              "spec": {
-                "pvc": {
-                  "storageClassName": "nfs-client",
-                  "accessModes": [
-                    "ReadWriteOnce"
-                  ],
-                  "resources": {
-                    "requests": {
-                      "storage": "10Gi"
-                    }
-                  }
-                },
-                "source": {
-                  "http": {
-                    "url": "https://download.fedoraproject.org/pub/fedora/linux/releases/41/Cloud/x86_64/images/Fedora-Cloud-Base-AmazonEC2-41-1.4.x86_64.raw.xz"
-                  }
-                }
-              }
-            }
-          ],
-          "running": true,
-          "template": {
-            "metadata": {
-              "labels": {
-                "kubevirt.io/vmpool": "test",
-                "kubevirt-manager.io/managed": "true",
-                "fw.kubevirt-manager.io/pool-name": "test",
-                "cloud-init.kubevirt-manager.io/username": ""
-              }
-            },
-            "spec": {
-              "domain": {
-                "devices": {
-                  "disks": [
-                    {
-                      "name": "disk1",
-                      "disk": {}
-                    },
-                    {
-                      "name": "disk3",
-                      "disk": {
-                        "bus": "virtio"
-                      }
-                    }
-                  ],
-                  "interfaces": [
-                    {
-                      "name": "net1",
-                      "bridge": {}
-                    }
-                  ],
-                  "networkInterfaceMultiqueue": true
-                },
-                "firmware": {
-                  "bootloader": {
-                    "bios": {}
-                  }
-                }
-              },
-              "priorityClassName": "vm-standard",
-              "networks": [
-                {
-                  "name": "net1",
-                  "pod": {}
-                }
-              ],
-              "volumes": [
-                {
-                  "name": "disk1",
-                  "dataVolume": {
-                    "name": "default-test-disk1"
-                  }
-                },
-                {
-                  "name": "disk3",
-                  "cloudInitNoCloud": {
-                    "userData": "#cloud-config\nmanage_etc_hosts: true\nuser: admin\npassword: admin\n",
-                    "networkData": "version: 1\nconfig:\n    - type: physical\n      name: enp1s0\n      subnets:\n      - type: dhcp\n"
-                  }
-                }
-              ]
-            }
-          },
-          "instancetype": {
-            "kind": "VirtualMachineClusterInstancetype",
-            "name": "test"
-          }
-        }
-      }
-    }
-  }
